Change Line
The goal is to remove the camlp4 while maintaining the same level of
convenience. The price to pay for this is a loss of some static
guarantees.

The old implementation featured a ['a Line.t] type, representing an
extensible record, where the type parameter described statically which
fields of the record were present, at any point of the program.

We switch to a [Line.t] type, which still describes an extensible
record, so that the expressivity remains the same and we can keep the
same APIs. However, presence or absence of fields cannot be check
statically as before (the underlying implementation for Line.t is a
[Mixmap.t], i.e. a map with values of heterogeneous types).